+++ internal/metric/repository.go
@@ -82,27 +82,3 @@ func (dr *DatabaseRepository) GetMetricByName(c context.Context, name string) (*
 	}
 	return &metric, nil
 }
-
-// MockRepository for testing
-type MockRepository struct {
-	Metrics []Metric
-}
-
-func NewMockRepository(metrics []Metric) *MockRepository {
-	return &MockRepository{
-		Metrics: metrics,
-	}
-}
-
-func (mr *MockRepository) GetMetrics(c context.Context) ([]Metric, error) {
-	return mr.Metrics, nil
-}
-
-func (mr *MockRepository) GetMetricByName(c context.Context, name string) (*Metric, error) {
-	for _, metric := range mr.Metrics {
-		if metric.Name == name {
-			return &metric, nil
-		}
-	}
-	return nil, errors.New("metric not found")
-}
